pyproject.toml
[project]
name = "runpod-wispr"
version = "0.1.0"
description = "RunPod serverless meeting transcription (MP3/MP4 → CSV with speaker ID)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
worker = [
    "runpod>=1.7.0",
    "whisperx>=3.1.0",
    "torch",
    "requests",
]
client = [
    "typer>=0.9.0",
    "google-cloud-storage>=2.0.0",
    "runpod>=1.7.0",
]

[dependency-groups]
client = [
    "typer>=0.9.0",
    "google-cloud-storage>=2.0.0",
    "runpod>=1.7.0",
]
worker = [
    "runpod>=1.7.0",
    "whisperx>=3.1.0",
    "torch",
    "requests",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
